Find records with valid external causes of injury icd-10-cm.

Find records with valid external causes of injury icd-10-cm.

matrix_valid_external(data, diag_ecode_col)

Arguments

data
input data

diag_ecode_col
column indices

Value

valid_external, a binary variable indicating whether the record has (value = 1) a valid external cause of injury icd-10-cm code

Examples


library(dplyr)
library(purrr)
set.seed(5)
icd10cm_data150 %>%
matrix_valid_external(diag_ecode_col = c(2:6)) %>%
sample_n(10)

#> # A tibble: 10 x 7
#>    uid   diagnosis_1 diagnosis_2 diagnosis_3 ecode1  ecode2 valid_external
#>    <chr> <chr>       <chr>       <chr>       <chr>   <chr>           <dbl>
#>  1 229   T405X1A     K920        I10         NA      NA                  1
#>  2 040   T40601A     G92         F319        Y92009  NA                  1
#>  3 017   T43292A     T43222A     T447X2A     Y92019  NA                  1
#>  4 245   T43592A     T43222A     T481X2A     X58XXXA NA                  1
#>  5 181   T43222A     R45851      R569        NA      NA                  1
#>  6 044   T438X2A     D61818      E11649      NA      NA                  1
#>  7 214   S270XXA     J9601       N179        W010XXA Y846                1
#>  8 194   T391X2A     G92         F3181       NA      NA                  1
#>  9 183   L03113      S61451A     I2510       W540XXA NA                  1
#> 10 202   T424X2A     G92         J9601       NA      NA                  1